@@ -1,19 +1,27 @@
 [workspace]
 resolver = "2"
 members = [
-  "crates/a2a-sdk",
-  "crates/core",
-  "crates/xds",
-  "crates/hbone",
-  "crates/proxy",
-  "crates/agentproxy",
-  "crates/agentgateway"
+    "crates/a2a-sdk",
+    "crates/agentgateway",
+    "crates/agentgateway-app",
+    "crates/core",
+    "crates/hbone",
+    "crates/mock-server",
+    "crates/xds",
+]
+default-members = [
+    "crates/a2a-sdk",
+    "crates/core",
+    "crates/xds",
+    "crates/hbone",
+    "crates/agentgateway",
+    "crates/agentgateway-app",
 ]
 
 [workspace.package]
 version = "0.4.12"
 edition = "2024"
-rust-version = "1.85"
+rust-version = "1.88"
 license = "Apache-2.0"
 publish = false
 
