@@ -1,7 +1,3 @@
-use agent_core::copy;
-use agent_core::prelude::*;
-use bytes::{BufMut, Bytes};
-use h2::Reason;
 use std::fmt::{Debug, Display};
 use std::hash::Hash;
 use std::io::Error;
@@ -11,6 +7,11 @@ use std::sync::Arc;
 use std::sync::atomic::{AtomicBool, AtomicU16, Ordering};
 use std::task::{Context, Poll};
 use std::time::Duration;
+
+use agent_core::copy;
+use agent_core::prelude::*;
+use bytes::{BufMut, Bytes};
+use h2::Reason;
 use tokio::sync::oneshot;
 use tracing::trace;
 
@@ -208,13 +209,10 @@ impl tokio::io::AsyncRead for TokioH2Stream {
 		copy::ResizeBufRead::poll_bytes(pinned, cx).map(|r| match r {
 			Ok(bytes) => {
 				if buf.remaining() < bytes.len() {
-					Err(Error::new(
-						std::io::ErrorKind::Other,
-						format!(
-							"kould overflow buffer of with {} remaining",
-							buf.remaining()
-						),
-					))
+					Err(Error::other(format!(
+						"kould overflow buffer of with {} remaining",
+						buf.remaining()
+					)))
 				} else {
 					buf.put(bytes);
 					Ok(())
@@ -350,6 +348,6 @@ fn h2_to_io_error(e: h2::Error) -> std::io::Error {
 	if e.is_io() {
 		e.into_io().unwrap()
 	} else {
-		std::io::Error::new(std::io::ErrorKind::Other, e)
+		std::io::Error::other(e)
 	}
 }
